Simplify Frontend Directory Structure"
