+++ frontend/app/_components/ui/text/wipetext/style.css
