+++ frontend/app/_components/ui/text/wipetext/wipetext.tsx
