+++ frontend/app/_features/AuthGuard.tsx
@@ -3,7 +3,7 @@
 import { usePathname, useRouter } from 'next/navigation';
 import { useSession } from 'next-auth/react';
 import { useEffect } from 'react';
-import { useTopPage } from '@/app/_features/top/TopPageContext';
+import { useTopPage } from '@/app/_contexts/TopPageContext';
 
 const unAuthenticatedPaths = ['/privacy-policy', '/terms-of-service'];
