+++ frontend/app/_types/Bookmark.ts
