+++ frontend/app/_types/Character.ts
