+++ frontend/app/_types/CharacterResponse.ts
