+++ frontend/app/_types/UserCharacter.ts
